Refactored to node for 0.4.0
