--- Tests/tests.html
@@ -1,29 +1,0 @@
-﻿<html>
-<head>
-    <meta charset="utf-8">
-    <title>Mocha Tests</title>
-    <link rel="stylesheet" href="../node_modules/mocha/mocha.css" />
-</head>
-<body>
-    <div id="mocha"></div>
-    <script src="../node_modules/mocha/mocha.js"></script>
-    <script src="../node_modules/chai/chai.js"></script>
-
-    <script src="../Source/ItemsHoldr.js"></script>
-
-    <script>
-        mocha.setup('bdd')
-        mocha.reporter('html');
-    </script>
-
-    <script src="tests.js"></script>
-
-    <script>
-        if (typeof mochaPhantomJS !== "undefined") {
-            mochaPhantomJS.run();
-        } else {
-            mocha.run();
-        }
-    </script>
-</body>
-</html>
