+++ test/unit/ItemValueTests/_.js
@@ -1,0 +1,9 @@
+define(["mocks"], function (mocks) {
+    return function () {
+        var expect = require("chai").expect;
+
+        it("tests haven't been implemented", function () {
+            // ...
+        });
+    };
+});
