+++ test/unit/index.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <title>EightBittr Unit Tests</title>
+        <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <link rel="stylesheet" href="../../node_modules/mocha/mocha.css" />
+    </head>
+    <body>
+        <div id="mocha"></div>
+
+        <script src="../../node_modules/chai/chai.js" type="text/javascript"></script>
+        <script src="../../node_modules/mocha/mocha.js" type="text/javascript"></script>
+        <script data-main="require" src="../../node_modules/requirejs/require.js" type="text/javascript"></script>
+
+        <script src="phantom.js" type="text/javascript"></script>
+    </body>
+</html>
