+++ test/unit/require.js
@@ -1,0 +1,5 @@
+require(
+    ["unit"],
+    function () {
+        mocha.run();
+    });
